Another example for duplicate @deprecated directive breaking the composition
This is for demonstration purposes. The absinthe schema will successfully
compile and the Phoenix server will start. However, when we attempt to
start the federated gateway, it will give us the following error:

```
IntrospectAndCompose failed to update supergraph with the following error: A valid schema couldn't be composed. The following composition errors were found:
        [products] The directive "@requires" can only be used once at this location.
Error: A valid schema couldn't be composed. The following composition errors were found:
        [products] The directive "@requires" can only be used once at this location.
```